Add tvOS framework
