--- Embrace.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module Embrace {
-  umbrella header "Embrace.h"
-
-  export *
-  module * { export * }
-}
